Revert "after doughnut dashboard and trying to add bar chart"
This reverts commit ef93f8f8f96c4a59c5f0b55670039480b14a10b5.
