+++ client/dark-web-scraper/package.json
@@ -13,10 +13,8 @@
     "@types/react-paginate": "^7.1.1",
     "axios": "^0.25.0",
     "bootstrap": "^5.1.3",
-    "chart.js": "^3.7.0",
     "react": "^17.0.2",
     "react-bootstrap": "^2.1.2",
-    "react-chartjs-2": "^4.0.1",
     "react-dom": "^17.0.2",
     "react-paginate": "^8.1.0",
     "react-scripts": "5.0.0",
