+++ client/dark-web-scraper/src/App.tsx
@@ -8,21 +8,11 @@ function App() {
   // const myContext = React.createContext(0)
   const [pastelist, setPasteList] = useState([]);
   const [search, setSearch] = useState([]);
-  const [stats, setStats] = useState({
-    General: 0,
-    Crypto: 0,
-    Hacking: 0,
-    DrugsAndWeapons: 0,
-    AdultsContent: 0,
-  });
   return (
     <div className="App">
       {/* <myContext.Provider value={900} > */}
       <Header pastelist={pastelist} setSearch={setSearch} />
       <PasteList
-      stats={stats}
-      // @ts-ignore
-      setStats={setStats}
         search={search}
         pastelist={pastelist}
         setPasteList={setPasteList}
