+++ client/dark-web-scraper/src/assets/styles/chat.css
@@ -1,0 +1,221 @@
+.chat-container {
+  overflow: hidden;
+  border: 1px solid black;
+  margin: 0 auto;
+  width: 90%;
+  max-width: 1200px;
+  border-radius: var(--radius);
+  background-color: var(--background);
+}
+
+.chat__box--container {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  max-height: 70vh;
+}
+
+.chat__box {
+  overflow: auto;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: reverse;
+      -ms-flex-direction: column-reverse;
+          flex-direction: column-reverse;
+}
+
+.chat__box--messages {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  gap: 1em;
+  padding: 1em;
+  background-image: var(--bg-chat-img);
+}
+
+.message__container--left {
+  -webkit-box-pack: start;
+      -ms-flex-pack: start;
+          justify-content: flex-start;
+}
+
+.message__container--right {
+  -webkit-box-pack: end;
+      -ms-flex-pack: end;
+          justify-content: flex-end;
+}
+
+.message--left {
+  border-radius: 0 20px 20px 20px;
+  background-color: var(--clr-primary);
+}
+
+.message--right {
+  border-radius: 20px 0 20px 20px;
+  background-color: var(--clr-accent);
+  color: var(--clr-light);
+}
+
+.message--right .message__user {
+  display: none;
+}
+
+.message__container {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+}
+
+.chat__message {
+  border: 1px solid black;
+  padding: .5em 1em;
+  max-width: 90%;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  gap: .3em;
+  overflow-wrap: break-word;
+}
+
+.chat__message .message__user {
+  font-weight: bold;
+  font-family: var(--font-title);
+}
+
+.chat__message .message__content {
+  padding: 0 .3em;
+}
+
+.chat__message .message__timestamp {
+  font-size: .7rem;
+  opacity: .75;
+}
+
+.send-message {
+  background-color: var(--clr-secondary);
+  padding: .5em;
+}
+
+.send-message .send-message__form {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  gap: .5em;
+  -webkit-box-pack: justify;
+      -ms-flex-pack: justify;
+          justify-content: space-between;
+}
+
+.send-message .send-message__input {
+  min-width: 0;
+  -webkit-box-flex: 1;
+      -ms-flex-positive: 1;
+          flex-grow: 1;
+  padding: .5em;
+  border: none;
+  border-radius: 30px;
+  background-color: var(--background);
+  color: inherit;
+  outline: none;
+}
+
+.send-message .send-message__button {
+  background-color: var(--clr-accent);
+  color: var(--clr-light);
+  padding: 0 1em;
+  border: none;
+  border-radius: 30px;
+  font-family: inherit;
+  font-weight: bold;
+  cursor: pointer;
+}
+
+.send-message .send-message__button .send-svg {
+  width: 23px;
+  aspect-ratio: 1;
+  vertical-align: bottom;
+}
+
+.send-message .send-message__button:hover,
+.send-message .send-message__button:focus {
+  opacity: .75;
+}
+
+.chat__aside {
+  overflow: auto;
+  max-height: 70vh;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  gap: 1px;
+}
+
+.chat__aside .aside__user {
+  padding: 1em;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  gap: .5em;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  background-color: var(--clr-primary);
+}
+
+.chat__aside .aside__user .online-circle {
+  border: 1px solid lime;
+  background-color: lime;
+  border-radius: 50px;
+  color: lime;
+  width: .7rem;
+  aspect-ratio: 1;
+}
+
+.chat__aside .aside__user .aside__user--email {
+  font-size: .7rem;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+/* 
+Screens 700px+
+*/
+@media (min-width: 700pX) {
+  .chat__message {
+    max-width: 70%;
+  }
+}
+
+/* 
+Screens 1000px+
+*/
+@media (min-width: 1000pX) {
+  .chat__message {
+    max-width: 45%;
+  }
+  .chat__aside {
+    width: 20%;
+  }
+  .chat__box--container {
+    width: 80%;
+  }
+}
+/*# sourceMappingURL=chat.css.map */
