+++ client/dark-web-scraper/src/assets/styles/footer.css
@@ -1,0 +1,52 @@
+/* 
+footer
+*/
+footer {
+  font-family: var(--font-text);
+  background-color: var(--background);
+  padding: 1em;
+  color: var(--foreground);
+  text-align: center;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
+}
+
+.copyright {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -ms-flex-wrap: wrap;
+      flex-wrap: wrap;
+  gap: 1em;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+}
+
+.copyright-link {
+  color: inherit;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  gap: .5em;
+  text-decoration: underline;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+}
+
+.copyright-link:hover,
+.copyright-link:focus {
+  opacity: .75;
+}
+
+.github-img {
+  width: 1.5em;
+  -webkit-filter: var(--inverted);
+          filter: var(--inverted);
+}
+/*# sourceMappingURL=footer.css.map */
