+++ client/dark-web-scraper/src/assets/styles/footer.scss
@@ -1,0 +1,34 @@
+/* 
+footer
+*/
+footer {
+    font-family: var(--font-text);
+    background-color: var(--background);
+    padding: 1em;
+    color: var(--foreground);
+    text-align: center;
+    display: flex;
+    justify-content: center;
+}
+.copyright {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 1em;
+    align-items: center;
+}
+.copyright-link {
+    color: inherit;
+    display: flex;
+    gap: .5em;
+    text-decoration: underline;
+    align-items: center;
+}
+.copyright-link:hover,
+.copyright-link:focus {
+    opacity: .75;
+}
+
+.github-img {
+    width: 1.5em;
+    filter: var(--inverted);
+}
