+++ client/dark-web-scraper/src/assets/styles/hello.css
@@ -1,0 +1,12 @@
+.hello {
+  background-color: var(--clr-accent);
+  color: var(--clr-light);
+  font-size: .9rem;
+}
+
+.hello-container {
+  margin: 0 auto;
+  width: 90%;
+  max-width: 1200px;
+}
+/*# sourceMappingURL=hello.css.map */
