+++ client/dark-web-scraper/src/assets/styles/hello.scss
@@ -1,0 +1,11 @@
+.hello {
+    background-color: var(--clr-accent);
+    color: var(--clr-light);
+    font-size: .9rem;
+}
+
+.hello-container {
+    margin: 0 auto;
+    width: 90%;
+    max-width: 1200px;
+}
