+++ client/dark-web-scraper/src/assets/styles/home.css
@@ -1,0 +1,4 @@
+.two-factor {
+  text-align: center;
+}
+/*# sourceMappingURL=home.css.map */
