+++ client/dark-web-scraper/src/assets/styles/home.scss
@@ -1,0 +1,3 @@
+.two-factor {
+  text-align: center;
+}
