+++ client/dark-web-scraper/src/assets/styles/login.scss
@@ -1,0 +1,9 @@
+/*  
+Log-in form
+*/
+.log-in-container {
+    width: 90%;
+    margin: 0 auto;
+    max-width: 400px;
+    padding: 3em 0;
+}
