+++ client/dark-web-scraper/src/assets/styles/sign-up.css
@@ -1,0 +1,10 @@
+/*  
+Sign-up form
+*/
+.sign-up-container {
+  width: 90%;
+  margin: 0 auto;
+  max-width: 400px;
+  padding: 3em 0;
+}
+/*# sourceMappingURL=sign-up.css.map */
