+++ client/dark-web-scraper/src/assets/styles/two-factor.scss
@@ -1,0 +1,14 @@
+/*  
+Log-in form
+*/
+.two-factor-container {
+    width: 90%;
+    margin: 0 auto;
+    max-width: 400px;
+    padding: 3em 0;
+}
+
+.qr-img {
+    display: block;
+    margin: 0 auto;
+}
