+++ client/dark-web-scraper/src/index.scss
@@ -117,7 +117,7 @@
     margin: 3px 10px;
     padding: 10px;
     box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
-     overflow-wrap: break-word;
+    overflow-wrap: break-word;
 }
 .pasteBoxBottom{
   display: flex;
@@ -177,11 +177,4 @@
   padding: 6px;
   background-color: var(--clr-accent-fade);
   border: 1px solid rgb(255, 255, 255);
-}
-
-
-// STATS
-.doughnutDiv{
-  width: 25%;
-  
 }
