+++ client/dark-web-scraper/src/index.tsx
@@ -10,4 +10,7 @@ ReactDOM.render(
   document.getElementById('root')
 );
 
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
