+++ scraper/src/app.ts
@@ -5,7 +5,7 @@ import cors from "cors";
 import morgan from "morgan";
 import morganHandler from "./middlewares/morgan";
 import errorHandler from "./middlewares/errorHandlers";
-
+import start from "./scraper";
 import listRouter from "./routes/listRoute";
 
 const app = express();
@@ -14,10 +14,10 @@ const app = express();
 app.use(express.json());
 app.use(cors());
 
-app.use(
-  morganHandler,
-  morgan(":method :url :status :res[content-length] - :response-time ms :body")
-);
+// app.use(
+//   morganHandler,
+//   morgan(":method :url :status :res[content-length] - :response-time ms :body")
+// );
